Change float32 to float16 to reduce model computation time

+++ raw_trainer/answer_tensor.py
@@ -6,4 +6,4 @@ settings = Settings("./settings.json")
 batch_size = settings.read("hyperparameters", "batch_size")
 num_groups = settings.read("data_info", "num_groups")
 
-answer_tensor = tf.placeholder(tf.float32, shape=[batch_size, num_groups])
+answer_tensor = tf.placeholder(tf.float16, shape=[batch_size, num_groups])
