+++ raw_trainer/input_tensor.py
@@ -5,4 +5,4 @@ settings = Settings("./settings.json")
 input_dimension = settings.read("data_info", "input_dimension")
 batch_size = settings.read("hyperparameters", "batch_size")
 
-input_tensor = tf.placeholder(tf.float32, shape=[None, input_dimension])
+input_tensor = tf.placeholder(tf.float16, shape=[None, input_dimension])
