+++ raw_trainer/neural_net/bincount_cnn.py
@@ -8,9 +8,9 @@ input_dimension = settings.read("data_info", "input_dimension")
 num_groups = settings.read("data_info", "num_groups")
 
 input_tensor_casted = tf.cast(input_tensor, tf.int32)
-bfd_casted = tf.map_fn(lambda t: tf.cast(tf.bincount(t, minlength=256, maxlength=256), tf.float32),
+bfd_casted = tf.map_fn(lambda t: tf.cast(tf.bincount(t, minlength=256, maxlength=256), tf.float16),
                        input_tensor_casted,
-                       dtype=tf.float32)
+                       dtype=tf.float16)
 bfd_reshaped = tf.reshape(bfd_casted, [-1, 16, 16, 1])
 
 w1 = tf.get_variable(name='w1', shape=[5, 5, 1, 2], initializer=tf.contrib.keras.initializers.he_normal())
